+++ src/IbanNet/Registry/IbanStructure.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
+using IbanNet.Registry.Patterns;
 using IbanNet.Validation;
 
 namespace IbanNet.Registry
