+++ src/IbanNet/Registry/Patterns/ITokenizer.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace IbanNet.Registry.Parsing
+namespace IbanNet.Registry.Patterns
 {
     /// <summary>
     /// Provides a way to classify sections of an input string into separate tokens.
