+++ src/IbanNet/Validation/Rules/IsMatchingStructureRule.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using IbanNet.Registry;
+using IbanNet.Registry.Patterns;
 using IbanNet.Validation.Results;
 
 namespace IbanNet.Validation.Rules
