+++ test/IbanNet.Tests/Registry/Patterns/PatternTests.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
-using IbanNet.Registry.Parsing;
 using Moq;
 using Xunit;
 
-namespace IbanNet.Registry
+namespace IbanNet.Registry.Patterns
 {
     public class PatternTests
     {
